refactor(mcpserver): extract MCP conversion engine into internal/mcpserver
Step 5 of protocolserver extraction (see .design/protocolserver.md):
move the MCP adapters/forwarders/loop-processor/stream-interceptor/
tool-executor/continuation-store (~3.6k lines, serving-side only) from
internal/server/module/mcp to internal/mcpserver. The admin surface
(handler.go/routes.go, /api/v1/mcp CRUD) stays in module/mcp.

This removes the last protocolserver -> internal/server/module/*
dependency. mcpserver has zero internal/server deps (only mcp/runtime,
mcp/tools, protocolserver/forwarding via context provider).
