+++ internal/mcpserver/anthropic_beta_adapter.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"encoding/json"
