+++ internal/mcpserver/anthropic_continuation_merge_test.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"testing"
