+++ internal/mcpserver/anthropic_v1_adapter.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"encoding/json"
