+++ internal/mcpserver/continuation_store.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"fmt"
