+++ internal/mcpserver/forwarder.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"context"
