+++ internal/mcpserver/generic_loop_processor.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"context"
