+++ internal/mcpserver/generic_stream_interceptor.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"encoding/json"
