+++ internal/mcpserver/message_extract.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"encoding/json"
