+++ internal/mcpserver/mixed_validation.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import "fmt"
