+++ internal/mcpserver/mixed_validation_test.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"testing"
