+++ internal/mcpserver/mock_adapter_test.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"context"
