+++ internal/mcpserver/openai_chat_adapter.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"fmt"
