+++ internal/mcpserver/tool_executor.go
@@ -1,4 +1,4 @@
-package mcp
+package mcpserver
 
 import (
 	"context"
